+++ src/ScopeProvider/index.ts
@@ -1,0 +1,1 @@
+export { ScopeContext, ScopeProvider } from './ScopeProvider';
